Add files via upload
The app.py file contains the flask implementation for the website as well as the LangGraph implementation which empowers the chatbot. The index.html file contains the layout for the website which acts as a user interface. 

The Langgraph_App.py file is the implementation of same LangGraph framework which empowers our chatbot. One can refer to it if the interface isn't important.